+++ torch_frame/__init__.py
@@ -1,7 +1,7 @@
 from .stype import (stype, numerical, categorical, text_embedded,
                     multicategorical)
 from .data import TensorFrame
-from .typing import TaskType, DataFrame, NAStrategy, TensorData
+from .typing import TaskType, DataFrame, NAStrategy
 from torch_frame.utils import save, load, cat  # noqa
 import torch_frame.data  # noqa
 import torch_frame.datasets  # noqa
@@ -20,7 +20,6 @@ __all__ = [
     'TaskType',
     'NAStrategy',
     'TensorFrame',
-    'TensorData',
     'save',
     'load',
     'cat',